+++ src/main/java/com/Chomusuke/gui/pane/TransactionPane.java
@@ -22,8 +22,8 @@ import com.chomusuke.gui.element.DateSelector;
 import com.chomusuke.gui.element.PlusButton;
 import com.chomusuke.gui.element.SquareButton;
 import com.chomusuke.gui.element.Tile.TransactionTile;
-import com.chomusuke.gui.stage.AddFileScreen;
-import com.chomusuke.gui.stage.AddTransactionScreen;
+import com.chomusuke.gui.popups.AddFileScreen;
+import com.chomusuke.gui.popups.AddTransactionScreen;
 import com.chomusuke.logic.Account;
 import com.chomusuke.logic.Storage;
 import com.chomusuke.logic.Transaction;
